Changes to the convolution function to avoid having to use EBImage. This should be faster as well.
NOTE: To use this, you need to change a few lines in script 3: in mine it is lines 524 and 530. Search for "adj.ag.f" – you need to replace:
    model_data_current <- adj.ag.f(m = model_data_current,
                                   r = rows,
                                   c = cols,
                                   f = adj_cell_mat,
                                   crop = "crop")
with:
    model_data_current <- adj.ag.f(m = model_data_current,
                                   r = rows,
                                   c = cols,
                                   crop = "crop")

i.e. to delete reference to "f", which we no longer need.

FYI, the "f" is the convolution filter. My re-write assumes that we're using a 3 x 3 matrix and so doesn't need this.